ref(models): divide models.py into individual models
The large models.py is rather cumbersome to work with.
