+++ tox.ini
@@ -82,7 +82,7 @@ warn_unused_ignores = True
 no_implicit_optional = True
 strict_equality = True
 
-[mypy-core.models]
+[mypy-core.models.*]
 warn_return_any = False
 
 [mypy-django.*]
